+++ docs/new-media.md
@@ -70,7 +70,7 @@ pageClass: routes
 
 ### 资讯热榜
 
-<Route author="nczitzk" example="/36kr/hot-list" path="/36kr/hot-list/:category?" :paramsDesc="['分类，默认为资讯人气榜']">
+<Route author="nczitzk" example="/36kr/hot-list" path="/36kr/hot-list/:category?" :paramsDesc="['分类，默认为24小时热榜']">
 
 | 24小时热榜 | 资讯人气榜 | 资讯综合榜 | 资讯收藏榜 |
 | ---------- | ---------- | ---------- | ---------- |
