+++ lib/v2/36kr/index.js
@@ -1,7 +1,8 @@
 const got = require('@/utils/got');
 const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
-const CryptoJS = require('crypto-js');
+
+const { rootUrl, ProcessItem } = require('./utils');
 
 const shortcuts = {
     '/information': '/information/web_news',
@@ -15,7 +16,6 @@ const shortcuts = {
 module.exports = async (ctx) => {
     const path = ctx.path.replace(/^\/news(?!flashes)/, '/information').replace(/\/search\/article/, '/search/articles');
 
-    const rootUrl = 'https://www.36kr.com';
     const currentUrl = `${rootUrl}${shortcuts.hasOwnProperty(path) ? shortcuts[path] : path}`;
 
     const response = await got({
@@ -42,34 +42,7 @@ module.exports = async (ctx) => {
         });
 
     if (!/^\/(search|newsflashes)/.test(path)) {
-        items = await Promise.all(
-            items.map((item) =>
-                ctx.cache.tryGet(item.link, async () => {
-                    const detailResponse = await got({
-                        method: 'get',
-                        url: item.link,
-                    });
-
-                    const cipherTextList = detailResponse.data.match(/{"state":"(.*)","isEncrypt":true}/) ?? [];
-                    if (cipherTextList.length !== 0) {
-                        const key = CryptoJS.enc.Utf8.parse('efabccee-b754-4c');
-                        const content = JSON.parse(
-                            CryptoJS.AES.decrypt(cipherTextList[1], key, {
-                                mode: CryptoJS.mode.ECB,
-                                padding: CryptoJS.pad.Pkcs7,
-                            })
-                                .toString(CryptoJS.enc.Utf8)
-                                .toString()
-                        ).articleDetail.articleDetailData.data;
-                        item.description = content.widgetContent;
-                    } else {
-                        const $ = cheerio.load(detailResponse.body);
-                        item.description = $('div.articleDetailContent').html();
-                    }
-                    return item;
-                })
-            )
-        );
+        items = await Promise.all(items.map((item) => ProcessItem(item, ctx.cache.tryGet)));
     }
 
     ctx.state.data = {
